+++ client/src/components/Sidebar.tsx
@@ -18,6 +18,7 @@ const Sidebar = () => {
             {isSidebarOpen && <Overlay onClick={() => dispatch(toggleSidebar())}/>}
             {isSidebarOpen && (
                 <SidebarContainer>
+                    <StyledLink to="/log">Login</StyledLink>
                     <StyledLink to="/">Home</StyledLink>
                     <StyledLink to="/">Anniversary</StyledLink>
                 </SidebarContainer>
