+++ client/src/index.tsx
@@ -9,7 +9,6 @@ import store from './store/index';
 import axios from "axios";
 
 axios.defaults.baseURL = 'http://localhost:8080';
-axios.defaults.headers.common['Content-Type'] = 'application/json';
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement
