+++ client/src/pages/Login.tsx
@@ -1,16 +1,58 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import kakaoLoginImage from '../assets/images/kakao_login.png';
 import axios from "axios";
 import tw from "tailwind-styled-components";
+import {useDispatch} from "react-redux";
+import {useLocation, useNavigate} from "react-router-dom";
 
 
 function Login() {
+    const dispatch = useDispatch();
+    const location = useLocation();
+    const queryParams = new URLSearchParams(location.search);
+    const navigate = useNavigate();
+
     const handleKakaoLoginBtn = () => {
-        axios.get('http://localhost:8080/oauth/kakao').then((res) => {
+        axios.get('api/v1/oauth/kakao').then((res) => {
             window.location.href = res.data;
         }).catch((err) => {
             console.log(err);
         });
+    }
+
+    useEffect(() => {
+        if (queryParams?.get("code")) {
+            const params = new URLSearchParams;
+            params.append('code', queryParams.get("code") || "");
+
+            axios.post('/api/v1/oauth/kakao/token', params)
+                .then((res) => {
+                    console.log('token', res.data);
+                    // dispatch(setAccessToken(res.data.access_token));
+                    sessionStorage.setItem('accessToken', res.data.access_token);
+                    // dispatch(setRefreshToken(res.data.refresh_token));
+                    handleGetUserInfo(res.data.access_token);
+                })
+                .catch((err) => {
+                    console.log(err);
+                })
+        }
+    }, []);
+
+    const handleGetUserInfo = (accessToken: string) => {
+        axios.post('/api/v1/oauth/kakao/access', accessToken, {
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        })
+            .then((res) => {
+                console.log('userInfo', res);
+                navigate('/');
+            })
+            .catch((err) => {
+                console.log(accessToken);
+                console.log(err);
+            })
     }
 
     return (
